% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_model.R
\name{build_confusion_mat}
\alias{build_confusion_mat}
\title{Build confusion matrix}
\usage{
build_confusion_mat(actual, predicted)
}
\arguments{
\item{actual}{Actual classes list}

\item{predicted}{Predicted classes list}
}
\value{
confusion.matrix Confusion matrix in table format.
}
\description{
Build a confusion matrix from actual and predicted class comparisons.
}
\examples{
 train.index <- split_data(iris.data, 0.7)
 training.data <- iris.data[train.index, ]
 test.data <- iris.data[-train.index, ]

 obj.func <- formula(Species ~ Sepal.Length + Sepal.Width)
 my.model <- train_tree(obj.func, training.data)

 predictions <- run_model(my.model, test.data)

 build_confusion_mat(predictions$Species, predictions$predict)

}

