% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris.data}
\alias{iris.data}
\title{Iris dataset}
\format{Data frame with 150 rows and 5 variables:
\describe{
 \item{Sepal.Length}{Sepal length, in cm}
 \item{Sepal.Width}{Sepal width, in cm}
 \item{Petal.Length}{Petal length, in cm}
 \item{Petal.Width}{Petal width, in cm}
}}
\source{
R Iris package
}
\usage{
iris.data
}
\description{
A dataset containing 150 measurements of iris flowers.
The target for modeling is Species.
}
\keyword{datasets}

