% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_model.R
\name{run_model}
\alias{run_model}
\title{Run the classification tree}
\usage{
run_model(model, new.data)
}
\arguments{
\item{model}{A model object trained with the \code{\link{train_tree}}
function of this package.}

\item{new.data}{Dataset to be classified using the defined model.}
}
\value{
predict Returns new.data dataset with classes appended
 as column "predict"
}
\description{
This function will execute a classification tree built with this
package on new data.
}
\examples{
 train.index <- split_data(iris.data, 0.7)
 training.data <- iris.data[train.index, ]
 test.data <- iris.data[-train.index, ]

 obj.func <- formula(Species ~ Sepal.Length + Sepal.Width)
 my.model <- train_tree(obj.func, training.data)

 predictions <- run_model(my.model, test.data)

}

