% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dataset.R
\name{split_data}
\alias{split_data}
\title{Splits train and test sets}
\usage{
split_data(data, split.pct)
}
\arguments{
\item{data}{Dataset for splitting, data frame.}

\item{split.pct}{The percentage of data to be used in training set, numeric between 0-1.}
}
\value{
train.set An index of records used in training, numeric list. Remove these
 indices for test set.
}
\description{
This function will split a dataset into training
and test groups. Group membership is defined by a
"train.set" index list.
}
\examples{
 train.index <- split_data(iris.data, 0.7)
 training.data <- iris.data[train.index, ]
 test.data <- iris.data[-train.index, ]

}

