% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_tree}
\alias{train_tree}
\title{Train a classification tree}
\usage{
train_tree(objective, data)
}
\arguments{
\item{objective}{The objective function for training, of type formula.}

\item{data}{Training dataset, a dataframe.}
}
\value{
tree.model A classification model.
}
\description{
This function trains a classification decision tree using the
\strong{rpart} package.
}
\examples{
 train.index <- split_data(iris.data, 0.7)
 training.data <- iris.data[train.index, ]
 test.data <- iris.data[-train.index, ]

 obj.func <- formula(Species ~ Sepal.Length + Sepal.Width)
 my.model <- train_tree(obj.func, training.data)

}

